fraud-ml-service/fraud-ml-service/requirements.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
xgboost
joblib
python-dotenv
